Convert remaining API routes.
